Update LCDTest.c
Decided against a keypad, took too much code to do decided to do the alarm hour and minute like the max charge capacity using incrementing buttons. When the alarm is to be set it creates a mini menu that allows the alarm hour to be set from 1 - 24. Then it will call Set_Alarm_Minute and will increment by 5 in stead of 1. The user will have to press set in order to move on to the next step of setting the alarm or will reprint the startup screen to display all the information to the user and give them the original settings back. There are still some adjustments to be made, i have to move everything down slightly again in order to fit the battery percentage at the top of the screen, or i could display it in the bottom left hand corner to save myself the trouble of moving everything down (which would take a long time). Still have to code in toggle alarm, but shouldn't take more than like 15 mins to do. Also want to add a line of code that adds a zero before the minute of the alarm if the alarm minutes is set to less than 10. Tbh its a lot of copy and paste but shouldn't take more than like 2 hrs to do. Took me a while to write this bit of code because i kept changing my mind of what i wanted to do.